cleanup, move common code to indicore
