+++ indidriverio.c
@@ -35,10 +35,10 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110 - 1301  USA
 #include <pthread.h>
 
 #include "indidriver.h"
+#include "indidevapi.h"
 #include "userio.h"
 #include "indiuserio.h"
 #include "indidriverio.h"
-
 
 
 /* Buffer size. Must be ^ 2 */
